sdk: simplify some helpers to improve codeclimate
